注册返现调整(谢雨奇)

+++ src/main/java/com/sysongy/poms/driver/service/impl/DriverServiceImpl.java
@@ -150,7 +150,7 @@ public class DriverServiceImpl implements DriverService {
     			List<SysCashBack> list=sysCashBackService.queryForBreak("201");
     			if (list!=null && list.size() > 0 ) {
     				SysCashBack back= list.get(0);//获取返现规则
-    				sysUserAccountService.addCashToAccount(record.getSysUserAccountId(), BigDecimal.valueOf(Long.valueOf(back.getCash_per())), GlobalConstant.OrderType.REGISTER_CASHBACK);
+    				sysUserAccountService.addCashToAccount(record.getSysUserAccountId(), BigDecimal.valueOf(Double.valueOf(back.getCash_per())), GlobalConstant.OrderType.REGISTER_CASHBACK);
 				}else{
 					logger.info("找不到匹配的返现规则，注册成功，返现失败");    
 				}
@@ -434,8 +434,8 @@ public class DriverServiceImpl implements DriverService {
 			
 			if (listBack!=null && listBack.size() > 0) {
 				SysCashBack back= listBack.get(0);//获取返现规则
-				sysUserAccountService.addCashToAccount(driver.getSysUserAccountId(), BigDecimal.valueOf(Long.valueOf(back.getThreshold_min_value())), GlobalConstant.OrderType.REGISTER_CASHBACK);
-		    	sysUserAccountService.addCashToAccount(invitation.getSysUserAccountId(), BigDecimal.valueOf(Long.valueOf(back.getThreshold_max_value())), GlobalConstant.OrderType.INVITED_CASHBACK);
+				sysUserAccountService.addCashToAccount(driver.getSysUserAccountId(), BigDecimal.valueOf(Double.valueOf(back.getThreshold_min_value())), GlobalConstant.OrderType.REGISTER_CASHBACK);
+		    	sysUserAccountService.addCashToAccount(invitation.getSysUserAccountId(), BigDecimal.valueOf(Double.valueOf(back.getThreshold_max_value())), GlobalConstant.OrderType.INVITED_CASHBACK);
 			}else{
 				logger.info("找不到匹配的返现规则，注册成功，返现失败");
 			}
